+++ casadi/solvers/dle_to_lr_dle.cpp
@@ -34,32 +34,32 @@
 
 #include <numeric>
 
-INPUTSCHEME(LR_DLEInput)
+INPUTSCHEME(DLEInput)
 
 using namespace std;
 namespace casadi {
 
   extern "C"
-  int CASADI_LRDLESOLVER_DLE_EXPORT
-  casadi_register_lrdlesolver_dle(LrDleInternal::Plugin* plugin) {
+  int CASADI_DLESOLVER_LRDLE_EXPORT
+  casadi_register_dlesolver_lrdle(DleInternal::Plugin* plugin) {
     plugin->creator = DleToLrDle::creator;
-    plugin->name = "dle";
+    plugin->name = "lrdle";
     plugin->doc = DleToLrDle::meta_doc.c_str();
     plugin->version = 21;
-    plugin->adaptorHasPlugin = DleSolver::hasPlugin;
+    plugin->adaptorHasPlugin = LrDleSolver::hasPlugin;
     return 0;
   }
 
   extern "C"
-  void CASADI_LRDLESOLVER_DLE_EXPORT casadi_load_lrdlesolver_dle() {
-    LrDleInternal::registerPlugin(casadi_register_lrdlesolver_dle);
+  void CASADI_DLESOLVER_LRDLE_EXPORT casadi_load_dlesolver_lrdle() {
+    DleInternal::registerPlugin(casadi_register_dlesolver_lrdle);
   }
 
   DleToLrDle::DleToLrDle(
-         const LrDleStructure& st) : LrDleInternal(st) {
+         const DleStructure& st) : DleInternal(st) {
 
     // set default options
-    setOption("name", "unnamed_dle_to_lr_dle"); // name of the function
+    setOption("name", "unnamed_lr_dle_to_dle"); // name of the function
 
     Adaptor::addOptions();
   }
@@ -70,36 +70,21 @@ namespace casadi {
 
   void DleToLrDle::init() {
     // Initialize the base classes
-    LrDleInternal::init();
+    DleInternal::init();
 
-    MX H = MX::sym("H", H_);
     MX A = MX::sym("A", A_);
-    MX C = MX::sym("C", C_);
     MX V = MX::sym("V", V_);
 
-    MX Vs = (V+V.T())/2;
-
-    MX CVC = mul(C, mul(V, C.T()));
-
-    // Create an DleSolver instance
-    solver_ = DleSolver(getOption(solvername()),
-                        dleStruct("a", A_, "v", CVC.sparsity()));
+    // Create an LrDleSolver instance
+    solver_ = LrDleSolver(getOption(solvername()),
+                          lrdleStruct("a", A_, "v", V_));
     if (hasSetOption(optionsname())) solver_.setOption(getOption(optionsname()));
     solver_.init();
 
-    std::vector<MX> Pr = solver_.call(dleIn("a", A, "v", CVC));
-    MX P = Pr[DLE_P];
+    std::vector<MX> Pr = solver_.call(lrdleIn("a", A, "v", V));
 
-    std::vector<MX> HPH(Hs_.size(), 0);
-    std::vector<MX> Hs = horzsplit(H, Hi_);
-    MX out = 0;
-
-    for (int k=0;k<Hs.size();++k) {
-      HPH[k] = mul(Hs[k].T(), mul(P, Hs[k]));
-    }
-
-    f_ = MXFunction(lrdpleIn("a", A, "v", V, "c", C, "h", H),
-                    lrdleOut("y", blkdiag(HPH)));
+    f_ = MXFunction(dleIn("a", A, "v", V),
+                    dleOut("p", Pr[DLE_P]));
     f_.init();
 
     Wrapper::checkDimensions();
@@ -112,12 +97,13 @@ namespace casadi {
 
   Function DleToLrDle::getDerivative(int nfwd, int nadj) {
     return f_.derivative(nfwd, nadj);
+
   }
 
 
   void DleToLrDle::deepCopyMembers(
       std::map<SharedObjectNode*, SharedObject>& already_copied) {
-    LrDleInternal::deepCopyMembers(already_copied);
+    DleInternal::deepCopyMembers(already_copied);
   }
 
   DleToLrDle* DleToLrDle::clone() const {
